requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers
Faker
pandas